fix(auth-interceptor): create interceptor to fix refresh errors
